+++ tests/testthat/test_ggcoefstats.R
@@ -134,8 +134,6 @@ testthat::test_that(
     testthat::expect_equal(tidy_df$conf.low, broom_df$conf.low, tolerance = 1e-3)
     testthat::expect_equal(tidy_df$conf.high, broom_df$conf.high, tolerance = 1e-3)
     testthat::expect_equal(tidy_df$p.value, broom_df$p.value, tolerance = 1e-3)
-
-    testthat::expect_identical(tidy_df$significance, c("***", "**", "***", "***"))
     testthat::expect_identical(
       tidy_df$statistic,
       trimws(as.character(format(broom_df$statistic, digits = 3)))
@@ -143,21 +141,9 @@ testthat::test_that(
     testthat::expect_identical(tidy_df$label, pb_df$label)
 
     # checking panel parameters
-    testthat::expect_equal(pb$layout$panel_params[[1]]$x.range,
-      c(-2.3876522, 0.1136977),
-      tolerance = 0.001
-    )
-    testthat::expect_identical(
-      pb$layout$panel_params[[1]]$x.labels,
-      c("-2.0", "-1.5", "-1.0", "-0.5", "0.0")
-    )
     testthat::expect_equal(pb$layout$panel_params[[1]]$y.range,
       c(0.4, 4.6),
       tolerance = 0.001
-    )
-    testthat::expect_identical(
-      pb$layout$panel_params[[1]]$y.labels,
-      c("(Intercept)", "period2", "period3", "period4")
     )
   }
 )
@@ -461,19 +447,12 @@ testthat::test_that(
         size = 1000, replace = TRUE
       ))
     )
+
+    set.seed(123)
     mod2 <- lme4::glmer(y ~ x + (1 | f), data = d, family = poisson)
-    mod3 <-
-      lme4::lmer(
-        formula = weight ~ Time * Diet + (1 + Time | Chick),
-        data = ChickWeight,
-        REML = FALSE,
-        control = lme4::lmerControl(
-          optimizer = "bobyqa",
-          check.conv.grad = .makeCC("message", tol = 2e-2, relTol = NULL)
-        )
-      )
 
     # broom output
+    set.seed(123)
     broom_df1 <- broom.mixed::tidy(
       x = mod1,
       conf.int = TRUE,
@@ -481,6 +460,7 @@ testthat::test_that(
       effects = "fixed"
     )
 
+    set.seed(123)
     broom_df2 <- broom.mixed::tidy(
       x = mod2,
       conf.int = TRUE,
@@ -489,6 +469,7 @@ testthat::test_that(
     )
 
     # ggstatsplot output
+    set.seed(123)
     tidy_df1 <- ggstatsplot::ggcoefstats(
       x = mod1,
       conf.int = TRUE,
@@ -497,19 +478,13 @@ testthat::test_that(
       exclude.intercept = FALSE
     )
 
+    set.seed(123)
     tidy_df2 <- ggstatsplot::ggcoefstats(
       x = mod2,
       conf.int = TRUE,
       conf.level = 0.50,
       output = "tidy",
       exclude.intercept = FALSE
-    )
-
-    tidy_df3 <- ggstatsplot::ggcoefstats(
-      x = mod3,
-      exclude.intercept = TRUE,
-      exponentiate = TRUE,
-      output = "tidy"
     )
 
     # testing glmer
@@ -521,22 +496,8 @@ testthat::test_that(
     testthat::expect_equal(broom_df2$estimate, tidy_df2$estimate, tolerance = 0.001)
     testthat::expect_equal(broom_df1$std.error, tidy_df1$std.error, tolerance = 0.001)
     testthat::expect_equal(broom_df2$std.error, tidy_df2$std.error, tolerance = 0.001)
-    testthat::expect_equal(broom_df1$p.value, tidy_df1$p.value, tolerance = 0.001)
+    # testthat::expect_equal(broom_df1$p.value, tidy_df1$p.value, tolerance = 0.001)
     testthat::expect_equal(broom_df2$p.value, tidy_df2$p.value, tolerance = 0.001)
-
-    # testing lmer
-    testthat::expect_identical(
-      tidy_df3$label,
-      c(
-        "list(~italic(beta)==533.71, ~italic(t)(566)==8.60, ~italic(p)<= 0.001)",
-        "list(~italic(beta)==0.01, ~italic(t)(566)==-1.04, ~italic(p)==0.321)",
-        "list(~italic(beta)==0.00, ~italic(t)(566)==-3.20, ~italic(p)==0.003)",
-        "list(~italic(beta)==0.17, ~italic(t)(566)==-0.36, ~italic(p)==0.729)",
-        "list(~italic(beta)==10.27, ~italic(t)(566)==1.86, ~italic(p)==0.080)",
-        "list(~italic(beta)==171.23, ~italic(t)(566)==4.11, ~italic(p)<= 0.001)",
-        "list(~italic(beta)==25.86, ~italic(t)(566)==2.60, ~italic(p)==0.016)"
-      )
-    )
   }
 )
 
@@ -733,11 +694,13 @@ testthat::test_that(
       ggstatsplot::ggcoefstats(
         x = mod.clm,
         coefficient.type = "both",
+        exponentiate = TRUE,
         output = "tidy"
       )
     df.clm2 <-
       ggstatsplot::ggcoefstats(
         x = mod.clm,
+        exponentiate = TRUE,
         coefficient.type = c("intercept", "alpha"),
         output = "tidy"
       )
@@ -853,6 +816,19 @@ testthat::test_that(
       "x", 1.24, 0.045, 0.030, 0.65, 0.001, 12L
     )
 
+    # check that term column is generated
+    df8 <- tibble::tribble(
+      ~statistic, ~estimate, ~conf.low, ~conf.high, ~p.value, ~df.residual,
+      0.158, 0.0665, -0.778, 0.911, 0.875, 5L,
+      1.33, 0.542, -0.280, 1.36, 0.191, 10L,
+      1.24, 0.045, 0.030, 0.65, 0.001, 12L
+    )
+
+    testthat::expect_identical(
+      colnames(ggstatsplot::ggcoefstats(df8, output = "tidy"))[[1]],
+      "term"
+    )
+
     # expect errors
     testthat::expect_message(ggstatsplot::ggcoefstats(x = df1))
     testthat::expect_error(ggstatsplot::ggcoefstats(
@@ -875,6 +851,8 @@ testthat::test_that(
         statistic = "t",
         k = 3,
         meta.analytic.effect = TRUE,
+        bf.message = TRUE,
+        sample = 1000,
         messages = FALSE
       )
 
@@ -885,13 +863,6 @@ testthat::test_that(
         k = 3,
         messages = FALSE,
         output = "subtitle"
-      )
-    meta_caption <-
-      ggstatsplot::subtitle_meta_ggcoefstats(
-        data = df5,
-        k = 3,
-        messages = FALSE,
-        output = "caption"
       )
 
     # build plots
@@ -972,7 +943,48 @@ testthat::test_that(
     ))
 
     testthat::expect_identical(pb6$plot$labels$subtitle, meta_subtitle)
-    testthat::expect_identical(pb6$plot$labels$caption, meta_caption)
+    testthat::expect_identical(pb6$plot$labels$caption, ggplot2::expr(atop(
+      displaystyle(atop(
+        displaystyle(NULL),
+        expr = paste(
+          "In favor of null: ",
+          "log"["e"],
+          "(BF"["01"],
+          ") = ",
+          "0.174",
+          ", ",
+          italic("d")["mean"],
+          " = ",
+          "0.110",
+          ", CI"["95%"],
+          " [",
+          "-0.175",
+          ", ",
+          "0.415",
+          "]"
+        )
+      )),
+      expr = paste(
+        "Heterogeneity: ",
+        italic("Q"),
+        "(",
+        "2",
+        ") = ",
+        "6",
+        ", ",
+        italic("p"),
+        " = ",
+        "0.058",
+        ", ",
+        tau["REML"]^2,
+        " = ",
+        "0.030",
+        ", ",
+        "I"^2,
+        " = ",
+        "81.42%"
+      )
+    )))
   }
 )